refactor: delete all deprecated attributes
