+++ analysis_options.yaml
@@ -2,8 +2,7 @@ include: package:lints/recommended.yaml
 analyzer:
   exclude: [
     "**/*.g.dart",
-    "**/*.freezed.dart",
-    "example/**",
+    "**/*.freezed.dart"
   ]
 linter:
   rules:
